Step 8: Router State and selector composition
