+++ src/products/products.module.ts
@@ -25,11 +25,11 @@ export const ROUTES: Routes = [
     component: fromContainers.ProductsComponent,
   },
   {
-    path: ':id',
+    path: 'new',
     component: fromContainers.ProductItemComponent,
   },
   {
-    path: 'new',
+    path: ':pizzaId',
     component: fromContainers.ProductItemComponent,
   },
 ];
