+++ src/products/store/index.ts
@@ -1,3 +1,4 @@
 export * from './reducers';
 export * from './actions';
 export * from './effects';
+export * from './selectors';
